методы работы с длительностью вынесены в common.js
